chore: created helper.go file and added validateUserInput() common functionality into it
