+++ main.go
@@ -11,14 +11,13 @@ var remainingTickets uint = 50
 var bookings = []string{}  // slice is dynamic in length, despite array is fixed in length
 
 func main() {
-	
 
 	greetUsers()	// function call
 
 	// fmt.Printf("conferenceTickets is %T, remainingTickets is %T, conferenceName is %T\n", conferenceTickets, remainingTickets, conferenceName)
 	
-
 	for {
+
 		firstName, lastName, email, userTickets := getUserInput()
 		isValidName, isValidEmail, isValidTicketNumber := validateUserInput(firstName, lastName, email, userTickets) // function call
 
@@ -79,13 +78,6 @@ func getFirstNames() []string {
 		firstNames = append(firstNames, names[0])
 	}
 	return firstNames
-}
-
-func validateUserInput(firstName string, lastName string, email string, userTickets uint) (bool, bool, bool) { // return values (bool, bool, bool)
-	isValidName := len(firstName) >= 2 && len(lastName) >= 2
-	isValidEmail := strings.Contains(email, "@")
-	isValidTicketNumber := userTickets > 0 && userTickets <= remainingTickets
-	return isValidName, isValidEmail, isValidTicketNumber // A Go function can return multiple values, here we return 3 bool values
 }
 
 func getUserInput() (string, string, string, uint){
